Minor comment tweak in Driver, larger comment tweak in WireTappedPrintStream, added some casts-N-stuff to PureListJUTest

+++ ProjectFortress/src/com/sun/fortress/interpreter/useful/PureListJUTest.java
@@ -26,7 +26,8 @@ public class PureListJUTest extends TestCase {
         int counter = 0;
         for (int elt : list) {
             counter++;
-            assertTrue("PureList iterator not returning correct elements", elt == array[elt]);
+            assertTrue("PureList iterator not returning correct elements",
+                    elt == ((Integer)array[elt]).intValue());
         }
         assertEquals("Incorrect number of iterations over list.", 4, counter);
     }
